+- modified and add feature catch up for contact module and chatterbox

+++ src/utils/contacts.py
@@ -12,7 +12,8 @@ from src.models.users import (
 )
 from src.models.mobile_numbers import (
     UserMobiles, UserMobilesSchema,
-    MobileNumbers
+    MobileNumbers, BlockedMobileNumbers,
+    BlockedMobileNumbersSchema
 )
 from src.models.organizations import (
     Organizations, OrganizationsSchema,
@@ -431,13 +432,13 @@ def save_user_ewi_restriction(restriction, user_id):
     """
     Function that save user ewi restriction
     """
+    if restriction != 0:
+        UserEwiRestrictions.query.filter(
+            UserEwiRestrictions.user_id == user_id).delete()
 
-    UserEwiRestrictions.query.filter(
-        UserEwiRestrictions.user_id == user_id).delete()
-
-    save_restriction_query = UserEwiRestrictions(
-        user_id=user_id, alert_level=restriction)
-    DB.session.add(save_restriction_query)
+        save_restriction_query = UserEwiRestrictions(
+            user_id=user_id, alert_level=restriction)
+        DB.session.add(save_restriction_query)
 
     return True
 
@@ -545,62 +546,47 @@ def remove_sites_with_ground_meas(
     """
     Function that remove site id with ground meas
     """
+    
     year = current_datetime.year
     month = current_datetime.month
     day = current_datetime.day
 
     routine_reminder_time = datetime(year, month, day, 9, 30)
-    end_time = datetime(year, month, day, 9, 35)
+    routine_end_time = datetime(year, month, day, 9, 35)
+    one_thirty_reminder_time = datetime(year, month, day, 13, 30)
+    one_thirty_end_time = datetime(year, month, day, 13, 35)
+    five_thirty_reminder_time = datetime(year, month, day, 5, 30)
+    five_thirty_end_time = datetime(year, month, day, 5, 35)
 
-    if routine_reminder_time < current_datetime < end_time:
+    if routine_reminder_time < current_datetime < routine_end_time:
         if routine_site_ids or extended_site_ids:
-            run_down_ts = routine_reminder_time - \
-                timedelta(hours=4, minutes=30)
-            mo_result = MarkerObservations.query.filter(
-                MarkerObservations.ts.between(run_down_ts, routine_reminder_time)).all()
+            result = get_site_with_observation_and_remove(routine_reminder_time, timedelta_hour=4)
 
-            for row in mo_result:
-                site_id = row.site_id
-                if site_id in routine_site_ids:
-                    routine_site_ids.remove(site_id)
-                if site_id in extended_site_ids:
-                    extended_site_ids.remove(site_id)
+            for site_id in result:
+                routine_site_ids.remove(site_id)
+                extended_site_ids.remove(site_id)
 
         if event_site_ids:
-            run_down_ts = routine_reminder_time - \
-                timedelta(hours=1, minutes=30)
-            mo_result = MarkerObservations.query.filter(
-                MarkerObservations.ts.between(run_down_ts, routine_reminder_time)).all()
+            result = get_site_with_observation_and_remove(routine_reminder_time, timedelta_hour=1)
 
-            for row in mo_result:
-                site_id = row.site_id
-                if site_id in event_site_ids:
-                    event_site_ids.remove(site_id)
+            for site_id in result:
+                event_site_ids.remove(site_id)
     else:
         routine_site_ids = []
         extended_site_ids = []
 
-        five_thirty_reminder_time = datetime(year, month, day, 5, 30)
-        five_thirty_end_time = datetime(year, month, day, 5, 35)
+        # reminder_time = five_thirty_reminder_time
 
         if five_thirty_reminder_time < current_datetime < five_thirty_end_time:
-            run_down_ts = five_thirty_reminder_time - \
-                timedelta(hours=1, minutes=30)
-
-        one_thirty_reminder_time = datetime(year, month, day, 13, 30)
-        one_thirty_end_time = datetime(year, month, day, 13, 35)
+            reminder_time = five_thirty_reminder_time
 
         if one_thirty_reminder_time < current_datetime < one_thirty_end_time:
-            run_down_ts = one_thirty_reminder_time - \
-                timedelta(hours=1, minutes=30)
+            reminder_time = one_thirty_reminder_time
+        
+        result = get_site_with_observation_and_remove(reminder_time, timedelta_hour=1)
 
-        mo_result = MarkerObservations.query.filter(
-            MarkerObservations.ts.between(run_down_ts, one_thirty_reminder_time)).all()
-
-        for row in mo_result:
-            site_id = row.site_id
-            if site_id in event_site_ids:
-                event_site_ids.remove(site_id)
+        for site_id in result:
+            event_site_ids.remove(site_id)
 
     final_site_ids = {
         "routine_site_ids": routine_site_ids,
@@ -608,10 +594,15 @@ def remove_sites_with_ground_meas(
         "extended_site_ids": extended_site_ids
     }
 
-    print(final_site_ids)
-
     return final_site_ids
 
+def get_site_with_observation_and_remove(reminder_time, timedelta_hour=1):
+    run_down_ts = reminder_time - \
+        timedelta(hours=timedelta_hour, minutes=30)
+    mo_result = MarkerObservations.query.with_entities(MarkerObservations.site_id).filter(
+        MarkerObservations.ts.between(run_down_ts, reminder_time)).all()
+
+    return mo_result
 
 def get_receipients_for_ground_meas(site_recipients):
     """
@@ -639,7 +630,7 @@ def get_receipients_for_ground_meas(site_recipients):
                 ).all()
 
             user_per_site_result = UsersRelationshipSchema(
-                many=True, exclude=["emails", "teams", "landline_numbers", "ewi_restriction"]
+                many=True, exclude=["emails", "teams", "landline_numbers", "ewi_restriction", "mobile_numbers.mobile_number.blocked_mobile"]
             ).dump(user_per_site_query).data
 
         feedback.append(
@@ -660,3 +651,13 @@ def get_site_ids(site_codes):
         site_ids.append(row.site_id)
 
     return site_ids
+    
+def get_blocked_numbers():
+    """
+    Function that gets blocked numbers
+    """
+    query = BlockedMobileNumbers.query.all()
+
+    result = BlockedMobileNumbersSchema(many=True).dump(query).data
+
+    return result
